feat: add multi-database support with feature flags and migration system
Implements comprehensive database abstraction layer supporting Firestore, PostgreSQL, and Redis with feature-flag controlled activation.

Includes repository base classes, connection pooling, retry logic, caching mixins, and complete migration framework with observability integration.

+++ src/infrastructure/observability/health_checker.py
@@ -36,7 +36,7 @@ class HealthCheckError(ApplicationError):
 class HealthChecker:
     """Component health monitoring system for application observability."""
 
-    def __init__(self, timeout: int) -> None:
+    def __init__(self, timeout: float) -> None:
         """Initialize health checker with configuration.
 
         Args:
